Move the compilation conventions to the shared project conventions
The idea is to consolidate the places where compilation is
affected, so that we don't have to guess where some flag is
being set.
For example, before this change, the atomicfu-applying logic
was split between these two files for no clear reason.
